Controlled accesses with Ability tokens

+++ app/Http/Controllers/Api/V1/TicketController.php
@@ -33,6 +33,12 @@ class TicketController extends ApiController
     {
         try {
             $user = User::findOrFail($request->input('data.relationships.author.data.id'));
+
+            //policy
+            Gate::authorize('create', null);
+
+            //TODO create ticket
+
         } catch (ModelNotFoundException $exception) {
             return $this->ok('User not found', [
                 'error' => 'The provided user id does not exist.'
@@ -92,6 +98,9 @@ class TicketController extends ApiController
         try {
             $ticket = Ticket::findOrFail($ticketId);
 
+            //policy
+            Gate::authorize('replace', $ticket);
+
             $ticket->update($request->mappedAttributes());
 
             return new TicketResource($ticket);
@@ -108,6 +117,10 @@ class TicketController extends ApiController
     {
         try {
             $ticket = Ticket::findOrFail($ticketId);
+
+            //policy
+            Gate::authorize('delete', $ticket);
+
             $ticket->delete();
 
             return $this->ok('Ticket successfully deleted');
